LayoutTests/imported/w3c/web-platform-tests/css/motion/animation/offset-rotate-interpolation-math-functions-expected.txt


FAIL CSS Transitions: property <offset-rotate> from [100deg] to [calc(sign(20rem - 20px) * 180deg)] at (-1) should be [20deg] assert_equals: expected "20deg " but got "200deg "
PASS CSS Transitions: property <offset-rotate> from [100deg] to [calc(sign(20rem - 20px) * 180deg)] at (0) should be [100deg]
FAIL CSS Transitions: property <offset-rotate> from [100deg] to [calc(sign(20rem - 20px) * 180deg)] at (0.125) should be [110deg] assert_equals: expected "110deg " but got "87.5deg "
FAIL CSS Transitions: property <offset-rotate> from [100deg] to [calc(sign(20rem - 20px) * 180deg)] at (0.875) should be [170deg] assert_equals: expected "170deg " but got "12.5deg "
FAIL CSS Transitions: property <offset-rotate> from [100deg] to [calc(sign(20rem - 20px) * 180deg)] at (1) should be [180deg] assert_equals: expected "180deg " but got "0deg "
FAIL CSS Transitions: property <offset-rotate> from [100deg] to [calc(sign(20rem - 20px) * 180deg)] at (2) should be [260deg] assert_equals: expected "260deg " but got "- 100deg "
FAIL CSS Transitions with transition: all: property <offset-rotate> from [100deg] to [calc(sign(20rem - 20px) * 180deg)] at (-1) should be [20deg] assert_equals: expected "20deg " but got "200deg "
PASS CSS Transitions with transition: all: property <offset-rotate> from [100deg] to [calc(sign(20rem - 20px) * 180deg)] at (0) should be [100deg]
FAIL CSS Transitions with transition: all: property <offset-rotate> from [100deg] to [calc(sign(20rem - 20px) * 180deg)] at (0.125) should be [110deg] assert_equals: expected "110deg " but got "87.5deg "
FAIL CSS Transitions with transition: all: property <offset-rotate> from [100deg] to [calc(sign(20rem - 20px) * 180deg)] at (0.875) should be [170deg] assert_equals: expected "170deg " but got "12.5deg "
FAIL CSS Transitions with transition: all: property <offset-rotate> from [100deg] to [calc(sign(20rem - 20px) * 180deg)] at (1) should be [180deg] assert_equals: expected "180deg " but got "0deg "
FAIL CSS Transitions with transition: all: property <offset-rotate> from [100deg] to [calc(sign(20rem - 20px) * 180deg)] at (2) should be [260deg] assert_equals: expected "260deg " but got "- 100deg "
FAIL CSS Animations: property <offset-rotate> from [100deg] to [calc(sign(20rem - 20px) * 180deg)] at (-1) should be [20deg] assert_equals: expected "20deg " but got "200deg "
PASS CSS Animations: property <offset-rotate> from [100deg] to [calc(sign(20rem - 20px) * 180deg)] at (0) should be [100deg]
FAIL CSS Animations: property <offset-rotate> from [100deg] to [calc(sign(20rem - 20px) * 180deg)] at (0.125) should be [110deg] assert_equals: expected "110deg " but got "87.5deg "
FAIL CSS Animations: property <offset-rotate> from [100deg] to [calc(sign(20rem - 20px) * 180deg)] at (0.875) should be [170deg] assert_equals: expected "170deg " but got "12.5deg "
FAIL CSS Animations: property <offset-rotate> from [100deg] to [calc(sign(20rem - 20px) * 180deg)] at (1) should be [180deg] assert_equals: expected "180deg " but got "0deg "
FAIL CSS Animations: property <offset-rotate> from [100deg] to [calc(sign(20rem - 20px) * 180deg)] at (2) should be [260deg] assert_equals: expected "260deg " but got "- 100deg "
FAIL Web Animations: property <offset-rotate> from [100deg] to [calc(sign(20rem - 20px) * 180deg)] at (-1) should be [20deg] assert_equals: expected "20deg " but got "200deg "
PASS Web Animations: property <offset-rotate> from [100deg] to [calc(sign(20rem - 20px) * 180deg)] at (0) should be [100deg]
FAIL Web Animations: property <offset-rotate> from [100deg] to [calc(sign(20rem - 20px) * 180deg)] at (0.125) should be [110deg] assert_equals: expected "110deg " but got "87.5deg "
FAIL Web Animations: property <offset-rotate> from [100deg] to [calc(sign(20rem - 20px) * 180deg)] at (0.875) should be [170deg] assert_equals: expected "170deg " but got "12.5deg "
FAIL Web Animations: property <offset-rotate> from [100deg] to [calc(sign(20rem - 20px) * 180deg)] at (1) should be [180deg] assert_equals: expected "180deg " but got "0deg "
FAIL Web Animations: property <offset-rotate> from [100deg] to [calc(sign(20rem - 20px) * 180deg)] at (2) should be [260deg] assert_equals: expected "260deg " but got "- 100deg "
FAIL CSS Transitions: property <offset-rotate> from [calc(sign(20rem - 20px) * 100deg)] to [calc(sign(20rem - 20px) * 180deg)] at (-1) should be [20deg] assert_equals: expected "20deg " but got "0deg "
FAIL CSS Transitions: property <offset-rotate> from [calc(sign(20rem - 20px) * 100deg)] to [calc(sign(20rem - 20px) * 180deg)] at (0) should be [100deg] assert_equals: expected "100deg " but got "0deg "
FAIL CSS Transitions: property <offset-rotate> from [calc(sign(20rem - 20px) * 100deg)] to [calc(sign(20rem - 20px) * 180deg)] at (0.125) should be [110deg] assert_equals: expected "110deg " but got "0deg "
FAIL CSS Transitions: property <offset-rotate> from [calc(sign(20rem - 20px) * 100deg)] to [calc(sign(20rem - 20px) * 180deg)] at (0.875) should be [170deg] assert_equals: expected "170deg " but got "0deg "
FAIL CSS Transitions: property <offset-rotate> from [calc(sign(20rem - 20px) * 100deg)] to [calc(sign(20rem - 20px) * 180deg)] at (1) should be [180deg] assert_equals: expected "180deg " but got "0deg "
FAIL CSS Transitions: property <offset-rotate> from [calc(sign(20rem - 20px) * 100deg)] to [calc(sign(20rem - 20px) * 180deg)] at (2) should be [260deg] assert_equals: expected "260deg " but got "0deg "
FAIL CSS Transitions with transition: all: property <offset-rotate> from [calc(sign(20rem - 20px) * 100deg)] to [calc(sign(20rem - 20px) * 180deg)] at (-1) should be [20deg] assert_equals: expected "20deg " but got "0deg "
FAIL CSS Transitions with transition: all: property <offset-rotate> from [calc(sign(20rem - 20px) * 100deg)] to [calc(sign(20rem - 20px) * 180deg)] at (0) should be [100deg] assert_equals: expected "100deg " but got "0deg "
FAIL CSS Transitions with transition: all: property <offset-rotate> from [calc(sign(20rem - 20px) * 100deg)] to [calc(sign(20rem - 20px) * 180deg)] at (0.125) should be [110deg] assert_equals: expected "110deg " but got "0deg "
FAIL CSS Transitions with transition: all: property <offset-rotate> from [calc(sign(20rem - 20px) * 100deg)] to [calc(sign(20rem - 20px) * 180deg)] at (0.875) should be [170deg] assert_equals: expected "170deg " but got "0deg "
FAIL CSS Transitions with transition: all: property <offset-rotate> from [calc(sign(20rem - 20px) * 100deg)] to [calc(sign(20rem - 20px) * 180deg)] at (1) should be [180deg] assert_equals: expected "180deg " but got "0deg "
FAIL CSS Transitions with transition: all: property <offset-rotate> from [calc(sign(20rem - 20px) * 100deg)] to [calc(sign(20rem - 20px) * 180deg)] at (2) should be [260deg] assert_equals: expected "260deg " but got "0deg "
FAIL CSS Animations: property <offset-rotate> from [calc(sign(20rem - 20px) * 100deg)] to [calc(sign(20rem - 20px) * 180deg)] at (-1) should be [20deg] assert_equals: expected "20deg " but got "0deg "
FAIL CSS Animations: property <offset-rotate> from [calc(sign(20rem - 20px) * 100deg)] to [calc(sign(20rem - 20px) * 180deg)] at (0) should be [100deg] assert_equals: expected "100deg " but got "0deg "
FAIL CSS Animations: property <offset-rotate> from [calc(sign(20rem - 20px) * 100deg)] to [calc(sign(20rem - 20px) * 180deg)] at (0.125) should be [110deg] assert_equals: expected "110deg " but got "0deg "
FAIL CSS Animations: property <offset-rotate> from [calc(sign(20rem - 20px) * 100deg)] to [calc(sign(20rem - 20px) * 180deg)] at (0.875) should be [170deg] assert_equals: expected "170deg " but got "0deg "
FAIL CSS Animations: property <offset-rotate> from [calc(sign(20rem - 20px) * 100deg)] to [calc(sign(20rem - 20px) * 180deg)] at (1) should be [180deg] assert_equals: expected "180deg " but got "0deg "
FAIL CSS Animations: property <offset-rotate> from [calc(sign(20rem - 20px) * 100deg)] to [calc(sign(20rem - 20px) * 180deg)] at (2) should be [260deg] assert_equals: expected "260deg " but got "0deg "
FAIL Web Animations: property <offset-rotate> from [calc(sign(20rem - 20px) * 100deg)] to [calc(sign(20rem - 20px) * 180deg)] at (-1) should be [20deg] assert_equals: expected "20deg " but got "0deg "
FAIL Web Animations: property <offset-rotate> from [calc(sign(20rem - 20px) * 100deg)] to [calc(sign(20rem - 20px) * 180deg)] at (0) should be [100deg] assert_equals: expected "100deg " but got "0deg "
FAIL Web Animations: property <offset-rotate> from [calc(sign(20rem - 20px) * 100deg)] to [calc(sign(20rem - 20px) * 180deg)] at (0.125) should be [110deg] assert_equals: expected "110deg " but got "0deg "
FAIL Web Animations: property <offset-rotate> from [calc(sign(20rem - 20px) * 100deg)] to [calc(sign(20rem - 20px) * 180deg)] at (0.875) should be [170deg] assert_equals: expected "170deg " but got "0deg "
FAIL Web Animations: property <offset-rotate> from [calc(sign(20rem - 20px) * 100deg)] to [calc(sign(20rem - 20px) * 180deg)] at (1) should be [180deg] assert_equals: expected "180deg " but got "0deg "
FAIL Web Animations: property <offset-rotate> from [calc(sign(20rem - 20px) * 100deg)] to [calc(sign(20rem - 20px) * 180deg)] at (2) should be [260deg] assert_equals: expected "260deg " but got "0deg "
